Create a transaction file and move relevant code there
